skyrl-sandbox/pyproject.toml
[project]
name = "skyrl-sandbox"
version = "0.1.0"
description = "Web interface for building SkyRL-Gym environments"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "jsonpath-ng>=1.6.0",
    "httpx>=0.26.0",
    "datasets>=2.18.0",
]

[project.scripts]
skyrl-sandbox = "skyrl_sandbox.app:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["skyrl_sandbox"]

[tool.setuptools.package-data]
skyrl_sandbox = ["templates/*", "static/*"]

